+++ src/HttpClient/middlewares/request/setupAxios/__tests__/axiosTracingTestSuite.ts
@@ -167,7 +167,7 @@ export const registerSharedTestSuite = (testSuiteConfig: TestSuiteConfig) => {
           expect((requestSpan as any)._logs[len - 1].fields[ErrorReportLogFields.ERROR_ID]).toBeDefined()
           expect((requestSpan as any)._logs[len - 1].fields[ErrorReportLogFields.ERROR_KIND]).toBeDefined()
           expect(
-            ((requestSpan as any)._logs[len - 1].fields.error.message as string).startsWith(
+            ((requestSpan as any)._logs[len - 1].fields[ErrorReportLogFields.ERROR_DETAILS].message as string).startsWith(
               testSuiteConfig.expects.error!.errorMessagePrefix
             )
           ).toBeTruthy()